Change event listener to popUp_element.js file
